feat: 后端常量定义迁移
- 将应用使用的部分常量迁移到utils/constants.rs以便于维护/复用
