+++ src-tauri/src/services/download/starter_installer_links.rs
@@ -9,8 +9,9 @@ use serde::Deserialize;
 // 1) 24 小时缓存读取/刷新 jar_lfs_links.json
 // 2) 解析 CNB 分层 JSON（types -> core -> version -> file/url）
 // 3) 按固定优先级选择安装器下载 URL
-pub const STARTER_INSTALLER_LINKS_URL: &str = "https://cnb.cool/SeaLantern-studio/ServerCore-Mirror/-/releases/download/26.02.27/jar_lfs_links.json";
-const STARTER_INSTALLER_LINKS_FILE: &str = "jar_lfs_links.json";
+
+///此处常量见 utils/costants.rs
+use crate::utils::constants::{STARTER_INSTALLER_LINKS_FILE, STARTER_INSTALLER_LINKS_URL};
 const STARTER_INSTALLER_LINKS_CACHE_TTL: Duration = Duration::from_secs(24 * 60 * 60);
 const STARTER_INSTALLER_FETCH_RETRY_LIMIT: usize = 3;
 const STARTER_INSTALLER_FETCH_RETRY_DELAY: Duration = Duration::from_secs(2);
