+++ src-tauri/src/services/server/manager.rs
@@ -11,8 +11,8 @@ use serde::{Deserialize, Serialize};
 use super::installer;
 use super::log_pipeline as server_log_pipeline;
 
-const DATA_FILE: &str = "sea_lantern_servers.json";
-const RUN_PATH_MAP_FILE: &str = "sea_lantern_run_path_map.json";
+///此处常量见 utils/costants.rs
+use crate::utils::constants::{DATA_FILE, RUN_PATH_MAP_FILE};
 
 /// 验证服务器名称，防止路径遍历攻击
 /// 返回清理后的名称或错误信息
