+++ src-tauri/src/utils/mod.rs
@@ -1,4 +1,5 @@
 pub mod cli;
+pub mod constants;
 pub mod downloader;
 pub mod logger;
 pub mod path;
